requirements.txt
nose==1.3.7
numpy==1.14.2
Pillow==5.0.0
progressbar2==3.36.0
python-utils==2.3.0
six==1.11.0
SQLAlchemy==1.2.5
